requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
dj-database-url==2.3.0
Django==5.1.2
django-allauth==65.1.0
django-js-asset==2.2.0
django-multiselectfield==0.1.12
gunicorn==23.0.0
humanize==4.11.0
idna==3.10
packaging==24.1
pillow==11.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
pytz==2024.2
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
whitenoise==6.7.0
